gtk+3: update to newer version to avoid an autoreconf that fails.
Autoreconf failed in some quite mysterious way and new version of GTK+3
don't have the issue so simply upgrade. However it cross-compiles an
executable that it intended to run on the build and therefore links with
the wrong flags compared to compilation. Fortunately it doesn't try to
run the executable so we can just fool make by touching the build
outputs.