icu4c: patch build system to have a sane versioning.
Icu4c has been outputting libraries as "libicuXXNN" where "XX" is the
subcomponent name and "NN" is the version number, i.e. they have been
putting the version number in the library name. Only on Windows.

Patch in order to remove the version number from there and put -1 at the
end of the dll name in order to be able to have install incompatible
versions side-by-side if the ABI were to ever break (so far it seems it
hasn't been the case).